Add optional shouldSkip() callback
The invocation context passed as `this` can be used by the optional
`shouldSkip()` callback to decide if the tracing of the specific hook
should be skipped.

If it returns `true` no other work is performed for the invocation,
including handling of arguments, and the `Interceptor` callbacks can
return early to maximize performance and minimize the risk of stability
issues.